Board + border: neatly handle other board sizes
The initial version had a variable which arbitrarily defined the radius
of a single hex based on the width of the screen. The position of most
things, including the border, was then determined based on this
variable. The result was that game could not handle board sizes other
than 11 by 11 (the default), since the hexagons could not change size.

Instead, begin with the placement of the border in a fixed location,
scaling the actual hexagons to fit neatly within the border depending on
the size of the board (bigger boards = smaller hexagons).