Put common stuff in main.py (non-functional)
Previously, the Game object created in main.py created its own Display.
Now both the Game and Display objects are created in main.py, and the
Display is passed to the Game. This eases the impending addition of a
menu (which needs a Display, but controls how the Game object should be
made...)

Also move the main game loop to main.py since it doesn't actually have
much to do with the game logic (it is mostly doing Pygame stuff).